moved resolved encryption code to dask-sidecar: only dask-sidecar shall know how to resolve
